[Player] the NVIDIA Shield workaround is back
thanks to Experience Upgrade 3.3 we are pleased to announce the return
of the famous Shield workaround. As already known the Shield starts stuttering
on video playback after 15 minutes on interlaced (1080i) content. Really well
done NVIDIA, ...

Yes, we already know how to tackle this. Just reset the codec before the nice
video experience starts (e.g. after 13 minutes).

This is just a quick and dirty version.